product-service-async/src/requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
sqlmodel==0.0.27
pydantic-settings==2.12.0
psycopg2-binary==2.9.11
asyncpg==0.31.0
pyyaml==6.0.3
prometheus_fastapi_instrumentator==7.1.0
